+++ src/simple_earn/rest_api/models/get_locked_redemption_record_response_rows_inner.rs
@@ -34,7 +34,7 @@ pub struct GetLockedRedemptionRecordResponseRowsInner {
     #[serde(rename = "type", skip_serializing_if = "Option::is_none")]
     pub r#type: Option<String>,
     #[serde(rename = "deliverDate", skip_serializing_if = "Option::is_none")]
-    pub deliver_date: Option<String>,
+    pub deliver_date: Option<i64>,
     #[serde(rename = "lossAmount", skip_serializing_if = "Option::is_none")]
     pub loss_amount: Option<String>,
     #[serde(rename = "isComplete", skip_serializing_if = "Option::is_none")]