+++ src/simple_earn/rest_api/models/get_locked_subscription_preview_response_inner.rs
@@ -35,18 +35,18 @@ pub struct GetLockedSubscriptionPreviewResponseInner {
     #[serde(rename = "nextPay", skip_serializing_if = "Option::is_none")]
     pub next_pay: Option<String>,
     #[serde(rename = "nextPayDate", skip_serializing_if = "Option::is_none")]
-    pub next_pay_date: Option<String>,
+    pub next_pay_date: Option<i64>,
     #[serde(rename = "valueDate", skip_serializing_if = "Option::is_none")]
-    pub value_date: Option<String>,
+    pub value_date: Option<i64>,
     #[serde(rename = "rewardsEndDate", skip_serializing_if = "Option::is_none")]
-    pub rewards_end_date: Option<String>,
+    pub rewards_end_date: Option<i64>,
     #[serde(rename = "deliverDate", skip_serializing_if = "Option::is_none")]
-    pub deliver_date: Option<String>,
+    pub deliver_date: Option<i64>,
     #[serde(
         rename = "nextSubscriptionDate",
         skip_serializing_if = "Option::is_none"
     )]
-    pub next_subscription_date: Option<String>,
+    pub next_subscription_date: Option<i64>,
 }
 
 impl GetLockedSubscriptionPreviewResponseInner {